# Conflicts:
#	src/main/scala/stockmonitoringbot/stockpriceservices/StockInfo.scala
#	src/main/scala/stockmonitoringbot/stockpriceservices/StockPriceService.scala
#	src/test/scala/stockmonitoringbot/stockpriceservices/AlphavantageStockPriceServiceTest.scala

+++ src/main/scala/stockmonitoringbot/datastorage/DataStorageWithStockInfoLoad.scala
@@ -19,8 +19,8 @@ trait DataStorageWithStockInfoLoad extends InMemoryDataStorage {
         super.getPrice(stock)
       else
         for {stockInfo <- getStockPriceInfo(stock)
-             _ <- addStock(stock, stockInfo.price)
-        } yield stockInfo.price
+             _ <- addStock(stock, stockInfo.price.toDouble)
+        } yield stockInfo.price.toDouble
     }
 
   override def addNotification(notification: Notification): Future[Unit] =
@@ -29,7 +29,7 @@ trait DataStorageWithStockInfoLoad extends InMemoryDataStorage {
         super.addNotification(notification)
       else
         for {stockInfo <- getStockPriceInfo(notification.stock)
-             _ <- addStock(notification.stock, stockInfo.price)
+             _ <- addStock(notification.stock, stockInfo.price.toDouble)
              _ <- super.addNotification(notification)
         } yield ()
     }
